+++ src/Wordpress/Template/Templates/PostListTable/Contracts/Factory.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace tiFy\Template\Templates\PostListTable\Contracts;
+namespace tiFy\Wordpress\Template\Templates\PostListTable\Contracts;
 
 use tiFy\Contracts\Template\{FactoryDb, TemplateFactory};
 use tiFy\Template\Templates\ListTable\Contracts\{Item as BaseItem, Factory as BaseFactory};
@@ -10,9 +10,9 @@ interface Factory extends BaseFactory
     /**
      * {@inheritDoc}
      *
-     * @return Db
+     * @return Db|null
      */
-    public function db(): FactoryDb;
+    public function db(): ?FactoryDb;
 
     /**
      * @inheritDoc