+++ src/Wordpress/Template/Templates/PostListTable/Contracts/Params.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace tiFy\Template\Templates\PostListTable\Contracts;
+namespace tiFy\Wordpress\Template\Templates\PostListTable\Contracts;
 
 use tiFy\Template\Templates\ListTable\Contracts\Params as BaseParams;